backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
pydantic>=2.0
pydantic-settings>=2.0
python-multipart>=0.0.9
numpy>=1.26
scipy>=1.13
scikit-learn>=1.5
biopython>=1.84
fpdf2>=2.8
